Move unmap methods to FileServiceIT
